requirements.txt
langgraph
langchain
redis
pydantic
python-dotenv
openai
google-generativeai
tqdm
psycopg2-binary
numpy
